+++ addon/mixins/reload-list-mixin.js
@@ -119,6 +119,7 @@ export default Ember.Mixin.create({
                             (filterPredicate ?
                               filterPredicate :
                               undefined));
+    this.set('resultPredicate', resultPredicate);
 
     if (resultPredicate) {
       builder.where(resultPredicate);