+++ CHANGELOG.md
@@ -4,10 +4,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-* Add `limitFunction` for export excel.
+* Add getting current `limitFunction` from data object's list when exporting data to Excel.
 
 ### Fixed
-* Fix export excel autorization error.
+* Fix autorization problems when exporting data to Excel.
+* `user-settings` service:
+    * Fix `userSettings` updating.
+* `flexberry-simpleolv` component:
+    * Fix columns width applying from `userSettings`.
+* `olv-toolbar` component:
+    * Fix `userSettings` names getting on first render.
 
 ## [0.9.0-beta.3] - 2017-03-29
 ### Added